soon I'll finish this shit

+++ ft_sultoa_base.c
@@ -6,7 +6,7 @@
 /*   By: dmaznyts <dmaznyts@student.unit.ua>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/16 11:55:46 by dmaznyts          #+#    #+#             */
-/*   Updated: 2017/08/22 17:28:35 by dmaznyts         ###   ########.fr       */
+/*   Updated: 2017/08/23 11:56:58 by dmaznyts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ static void	f(unsigned long value, int base, char *str, int *i)
 	char	*tmp;
 
 	tmp = "0123456789abcdef";
-	if (value >= base)
+	if (value >= (unsigned long)base)
 		f(value / base, base, str, i);
 	str[(*i)++] = tmp[value % base];
 }
